requirements.txt
argcomplete==1.9.3
boto3==1.9.248
botocore==1.12.248
certifi==2019.9.11
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
docutils==0.15.2
durationpy==0.5
ecdsa==0.13.3
Flask==1.1.1
Flask-Login==0.4.1
future==0.16.0
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.1.1
placebo==0.9.0
pyasn1==0.4.7
python-dateutil==2.6.1
python-dotenv==0.10.3
python-jose==3.0.1
python-slugify==1.2.4
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.2
Unidecode==1.1.1
urllib3==1.25.6
Werkzeug==0.16.0
wsgi-request-logger==0.4.6
zappa==0.48.2
